moved alert component ui into our own folder
