+++ components/Alert/Alert.tsx
@@ -1,41 +1,59 @@
-import { Body, Description, Title } from '@/components/ui/alert';
-import { FcApproval, FcCancel, FcHighPriority, FcSms } from 'react-icons/fc';
+import * as React from 'react';
+import { cva, type VariantProps } from 'class-variance-authority';
 
-interface AlertProps {
-  variant: 'success' | 'error' | 'normal' | 'warning';
-}
+import { cn } from '@/utils/utils';
 
-const variantConfig = {
-  success: {
-    title: 'Success!',
-    description: 'Your action was successful.',
-    icon: FcApproval,
+const alertVariants = cva(
+  'relative w-full rounded-lg border p-4 [&>svg~*]:pl-7 [&>svg+div]:translate-y-[-3px] [&>svg]:absolute [&>svg]:left-4 [&>svg]:top-4 [&>svg]:text-foreground',
+  {
+    variants: {
+      variant: {
+        default: 'bg-background text-foreground',
+        destructive:
+          'border-destructive/50 text-destructive dark:border-destructive [&>svg]:text-destructive',
+      },
+    },
+    defaultVariants: {
+      variant: 'default',
+    },
   },
-  error: {
-    title: 'Error!',
-    description: 'There was an error processing your request.',
-    icon: FcHighPriority,
-  },
-  normal: {
-    title: 'Normal!',
-    description: 'Here is some important information.',
-    icon: FcSms,
-  },
-  warning: {
-    title: 'Warning!',
-    description: 'Please be cautious about this.',
-    icon: FcCancel,
-  },
-};
+);
 
-export default function Alert({ variant }: AlertProps) {
-  const { title, description, icon: Icon } = variantConfig[variant];
+const Body = React.forwardRef<
+  HTMLDivElement,
+  React.HTMLAttributes<HTMLDivElement> & VariantProps<typeof alertVariants>
+>(({ className, variant, ...props }, ref) => (
+  <div
+    ref={ref}
+    role="alert"
+    className={cn(alertVariants({ variant }), className)}
+    {...props}
+  />
+));
+Body.displayName = 'Alert';
 
-  return (
-    <Body className="mt-2 w-80 rounded p-4">
-      {Icon && <Icon className="h-4 w-4" />}
-      <Title>{title}</Title>
-      <Description>{description}</Description>
-    </Body>
-  );
-}
+const Title = React.forwardRef<
+  HTMLParagraphElement,
+  React.HTMLAttributes<HTMLHeadingElement>
+>(({ className, ...props }, ref) => (
+  <h5
+    ref={ref}
+    className={cn('mb-1 font-medium leading-none tracking-tight', className)}
+    {...props}
+  />
+));
+Title.displayName = 'AlertTitle';
+
+const Description = React.forwardRef<
+  HTMLParagraphElement,
+  React.HTMLAttributes<HTMLParagraphElement>
+>(({ className, ...props }, ref) => (
+  <div
+    ref={ref}
+    className={cn('text-sm [&_p]:leading-relaxed', className)}
+    {...props}
+  />
+));
+Description.displayName = 'AlertDescription';
+
+export { Body, Title, Description };
